==> notebooks/preprocessing/2-calendar_data_compile.ipynb <==
from pathlib import Path
from datetime import datetime
from os import PathLike
from typing import Union

# data manipulation and analysis
import numpy as np 
import pandas as pd

# Visualization
import matplotlib.pyplot as plt
from matplotlib.dates import DateFormatter
import matplotlib.dates as mdates
from matplotlib.dates import date2num
import seaborn as sns
from prettytable import PrettyTable
import plotly.express as px # An interactive graphing library that makes interactive, publication-quality graphs online.
import plotly.graph_objs as go # An interactive graphing library that makes interactive, publication-quality graphs online.

# ML model selection and evaluation
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score, KFold
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix

# preprocessing
from sklearn.preprocessing import LabelEncoder, StandardScaler

pd.set_option("display.max_rows", None, "display.max_columns", None)

imputed_demand_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/research/data/3-imputed/demand'
imputed_demand_path = Path(imputed_demand_dir)

calculated_features_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/research/data/4-calculated-features'
calculated_features_path = Path(calculated_features_dir)

print(imputed_demand_path)

demand_df = pd.read_csv(imputed_demand_path/'demand.csv')

demand_df.info()

features_df = demand_df.copy(deep=True)
features_df.head()

features_df['Unnamed: 0'] = pd.to_datetime(features_df['Unnamed: 0'])

features_df.set_index('Unnamed: 0', inplace=True)

features_df['hour_of_day'] = features_df.index.hour
features_df['year'] = features_df.index.year
features_df['month'] = features_df.index.month
features_df['day_of_week'] = features_df.index.dayofweek
features_df['day_of_year'] = features_df.index.dayofyear
# features_df['week_of_year'] = features_df.index.weekofyear
features_df['week_of_year'] = features_df.index.isocalendar().week
features_df['quarter'] = features_df.index.quarter
features_df.drop(columns=['ont_demand'], inplace=True)

features_df.head()

import holidays


hols = holidays.Canada(state='ON') # default is ontario Holidays
print(features_df.loc['2018-01-01'].index.date[0] in hols)
print(features_df.loc['2018-12-27'].index.date[0] in hols)

# adding column if each day is holiday or not
features_df['stat_hol'] = pd.Series(features_df.index.date).apply(lambda x: x in hols).values
features_df.head()

features_df.tail()

from astral import Astral
import datetime
a = Astral()
city_name='Toronto'
city = a[city_name]
#city.latitude
date = datetime.date(2019, 7, 2)
sun = city.sun(date=date, local = True)

print(sun['sunrise'])
print(sun['sunset'])
print(type(sun['sunrise']))

print(features_df.loc['2018-01-01'].index[0])
print(features_df.loc['2018-12-27'].index[0])

features_df.head()

def get_daylight_hours(row, city):
    sun = city.sun(date=row.name, local=True)
    sunrise = sun['sunrise'].replace(tzinfo=None) ; sunset = sun['sunset'].replace(tzinfo=None)
    bool_val = (row.name > sunrise) & (row.name < sunset)
    return bool_val


a = Astral()
city = a['Toronto']
features_df['day_light_hours'] = features_df.apply(get_daylight_hours, city=city, axis=1)
features_df.head()

features_df.tail()

features_df.info()

features_df.to_csv(calculated_features_path / 'calendar.csv')
