==> notebooks/preprocessing/4-demand-weather-merge.ipynb <==
from pathlib import Path
from datetime import datetime
from os import PathLike
from typing import Union
import dateutil

# data manipulation and analysis
import numpy as np 
import pandas as pd
from astral import Astral

# Visualization
import matplotlib.pyplot as plt
from matplotlib.dates import DateFormatter
import matplotlib.dates as mdates
from matplotlib.dates import date2num
import seaborn as sns
from prettytable import PrettyTable
import plotly.express as px # An interactive graphing library that makes interactive, publication-quality graphs online.
import plotly.graph_objs as go # An interactive graphing library that makes interactive, publication-quality graphs online.

# ML model selection and evaluation
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score, KFold
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix

# preprocessing
from sklearn.preprocessing import LabelEncoder, StandardScaler

pd.set_option("display.max_rows", None, "display.max_columns", None)

imputed_demand_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/resrc/data/3-imputed/demand/demand.csv'
imputed_demand_path = Path(imputed_demand_dir)

imputed_weather_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/resrc/data/3-imputed/weather-toronto'
imputed_weather_path = Path(imputed_weather_dir)

calculated_features_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/resrc/data/4-calculated-features'
calculated_features_path = Path(calculated_features_dir)

interim_data_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/resrc/data/5-interim'
interim_data_path = Path(interim_data_dir)

clean_data_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/resrc/data/6-clean'
clean_data_path = Path(clean_data_dir)

### Join calendar features and imputed demand data


cal_df = pd.read_csv(calculated_features_path / 'calendar.csv', index_col=0, parse_dates=True,date_parser=dateutil.parser.parse)
cal_df.info()

power_df = pd.read_csv(imputed_demand_path / 'demand.csv', index_col=0, parse_dates=True,
                      date_parser=dateutil.parser.parse)
power_df.rename(columns={'ont_demand': 'hourly_demand'}, inplace=True)

power_df.info()

feat_df = pd.concat([cal_df, power_df], axis=1, sort=True)
feat_df.head()

feat_df.shape

feat_df.info()

feat_df.to_csv(interim_data_path / 'demand.csv')

### Join demand, demand features, calendar features, weather features

int_df = pd.read_csv(interim_data_path / 'demand.csv', index_col=0, parse_dates=True,date_parser=dateutil.parser.parse)
int_df.rename(columns ={'ont_demand': 'hourly_demand'}, inplace=True)
int_df.info()

int_df.head()

int_df.shape

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,'hmdx': np.float64, 'wind_chill': np.float64, 'weather': object}
weather_df = pd.read_csv(imputed_weather_path / "weather_toronto.csv", parse_dates=True,dtype=dtypes, index_col=0, infer_datetime_format=True)

weather_df.info()

weather_df.head()

weather_df.shape

clean_df = pd.concat([weather_df, int_df], axis=1, sort=True)
clean_df.head()

clean_df.shape

clean_df.isnull().sum()

# Create a daily_peak column
# Peak Daily demand
clean_df = clean_df.join(feat_df.groupby(by=['year', 'day_of_year'])['hourly_demand'].max(),on=['year', 'day_of_year'], rsuffix='_peak_in_day_in_year')
clean_df.rename(columns={'hourly_demand_peak_in_day_in_year':'daily_peak'}, inplace=True)

clean_df.to_csv(clean_data_path / 'oversized.csv')

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,'hmdx': np.float64, 'wind_chill': np.float64, 'weather': object,'hour_of_day': np.float64, 'year': np.float64, 'month': np.float64,'day_of_week': np.float64, 'day_of_year': np.float64, 'week_of_year': np.float64,'quarter': np.float64, 'stat_hol': object, 'day_light_hours': object,'hourly_demand': np.float64, 'daily_peak': np.float64, 'weekly_peak': np.float64,'monthly_peak': np.float64, 'quarterly_peak': np.float64, 'annual_peak': np.float64,'hour_as_pct_of_daily_peak': np.float64,'daily_peak_as_pct_of_weekly_peak': np.float64,'monthly_peak_as_pct_of_annual_peak': np.float64,
'quarterly_peak_as_pct_of_annual_peak': np.float64, 'daily_peak_as_pct_of_annual_peak': np.float64}

oversized_df = pd.read_csv(clean_data_path / "oversized.csv", parse_dates=True,dtype=dtypes, index_col=0, infer_datetime_format=True)

oversized_df.info()

clean_df = oversized_df.dropna(subset=['hourly_demand'])

clean_df.to_csv(clean_data_path / "clean.csv")

clean_df.tail()

clean_df.isnull().sum()

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,'hmdx': np.float64, 'wind_chill': np.float64, 'weather': object,'hour_of_day': np.float64, 'year': np.float64, 'month': np.float64,'week_of_year': np.float64,'quarter': np.float64, 'stat_hol': object, 'day_light_hours': object,'weekly_peak': np.float64,'annual_peak': np.float64,'daily_peak_as_pct_of_weekly_peak': np.float64,'monthly_peak_as_pct_of_annual_peak': np.float64,'daily_peak_as_pct_of_annual_peak': np.float64}

df = pd.read_csv(clean_data_path / "clean.csv", parse_dates=True, dtype=dtypes, index_col=0, infer_datetime_format=True)

df.head()

clean_cut_df = clean_df.loc['1994': '2018']
clean_cut_df.to_csv(clean_data_path / "clean-cut.csv")

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,
          'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,
          'hmdx': np.float64, 'wind_chill': np.float64, 'weather': object,
         'hour_of_day': np.float64, 'year': np.float64, 'month': np.float64,
         'day_of_week': np.float64, 'day_of_year': np.float64, 'week_of_year': np.float64,
          'quarter': np.float64, 'stat_hol': object, 'day_light_hours': object,
          'hourly_demand': np.float64, 'daily_peak': np.float64, 'weekly_peak': np.float64,
          'monthly_peak': np.float64, 'quarterly_peak': np.float64, 'annual_peak': np.float64,
          'hour_as_pct_of_daily_peak': np.float64, 'daily_peak_as_pct_of_weekly_peak': np.float64,
         'monthly_peak_as_pct_of_annual_peak': np.float64,
          'quarterly_peak_as_pct_of_annual_peak': np.float64, 'daily_peak_as_pct_of_annual_peak': np.float64}

clean_cut_df = pd.read_csv(clean_data_path / "clean-cut.csv", parse_dates=True,
                 dtype=dtypes, index_col=0, infer_datetime_format=True)

clean_cut_df.info()

clean_cut_df.head()

clean_cut_df.tail()

clean_cut_df.isnull().sum()

clean_cut_df.shape

clean_cut_df['weather'].nunique()
