==> notebooks/preprocessing/6-to_daily_data.ipynb <==
from pathlib import Path
from datetime import datetime
from os import PathLike
from typing import Union
import dateutil
import math

# data manipulation and analysis
import numpy as np 
import pandas as pd
from astral import Astral

# Visualization
import matplotlib.pyplot as plt
from matplotlib.dates import DateFormatter
import matplotlib.dates as mdates
from matplotlib.dates import date2num
import seaborn as sns
from prettytable import PrettyTable
import plotly.express as px # An interactive graphing library that makes interactive, publication-quality graphs online.
import plotly.graph_objs as go # An interactive graphing library that makes interactive, publication-quality graphs online.

# ML model selection and evaluation
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score, KFold
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix

# preprocessing
from sklearn.preprocessing import LabelEncoder, StandardScaler

pd.set_option("display.max_rows", None, "display.max_columns", None)

clean_data_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/research/data/6-clean'
clean_data_path = Path(clean_data_dir)

usable_data_dir = '/Users/yashwanthkaruparthi/Developer/energy_demand/research/data/7-usable'
usable_data_path = Path(usable_data_dir)

df_hour = pd.read_csv(clean_data_path / 'for_todaily_conversion.csv')

df_hour['datetime'] = pd.to_datetime(df_hour['datetime'])

df_hour['date'] = df_hour['datetime'].dt.date

df_hour.head()

df_hour.drop(columns = 'weather', inplace=True)

df_hour.head()

# df_hour.set_index(['datetime'], inplace = True)
df_hour.head()

df_hour.columns

# Define the aggregation methods for each type of column
aggregation_methods = {
    'temp': 'mean',
    'dew_point_temp': 'mean',
    'rel_hum': 'mean',
    'wind_speed': 'mean',
    'visibility': 'mean',
    'press': 'mean',
    'wind_chill': 'mean',
    'hourly_demand': 'mean',
    'hmdxx': 'mean',
    'temp_min': 'min',
    'temp_max': 'max',
    'temp_mean': 'mean',
    'temp_median': 'median',
    'dew_point_temp_min': 'min',
    'dew_point_temp_max': 'max',
    'dew_point_temp_mean': 'mean',
    'dew_point_temp_median': 'median',
    'rel_hum_min': 'min',
    'rel_hum_max': 'max',
    'rel_hum_mean': 'mean',
    'rel_hum_median': 'median',
    'visibility_min': 'min',
    'visibility_max': 'max',
    'visibility_mean': 'mean',
    'visibility_median': 'median',
    'press_min': 'min',
    'press_max': 'max',
    'press_mean': 'mean',
    'press_median': 'median',
    'hmdxx_min': 'min',
    'hmdxx_max': 'max',
    'hmdxx_mean': 'mean',
    'hmdxx_median': 'median'
}

# Group by date and aggregate
daily_data = df_hour.groupby('date').agg(aggregation_methods).reset_index()


daily_data.head()

In this dataset, `'day_of_week', 'day_of_year', 'week_of_year', 'quarter'` are removed. If necessary for improving model performance add these

daily_data.shape

daily_data.tail()

unique_dates_count = daily_data['date'].nunique()
total_rows = len(daily_data)

date_range = (daily_data['date'].min(), daily_data['date'].max())

original_summary = df_hour[['temp', 'dew_point_temp', 'rel_hum', 'wind_speed', 'visibility', 'press']].describe()
daily_summary = daily_data[['temp', 'dew_point_temp', 'rel_hum', 'wind_speed', 'visibility', 'press']].describe()

unique_dates_count, total_rows, date_range, original_summary, daily_summary


daily_data.to_csv(usable_data_path / '1-hourly_data.csv')
